docs(modal): rename example to so that it gets displayed correctly in stackblitz
