+++ packages/angular-test-app/src/preview-examples/modal-by-instance.ts
@@ -12,7 +12,7 @@ import { ModalService } from '@siemens/ix-angular';
 import ModalByInstanceExample from './modal-by-instance-content';
 
 @Component({
-  selector: 'app-example',
+  selector: 'app-example-instance',
   template: '<ix-button (click)="openModal()">Show modal</ix-button>',
 })
 export default class ModalByInstance {
