+++ packages/angular-test-app/src/preview-examples/modal-by-template.ts
